+++ packages/inprodi-core/src/RHFPasswordInput/RHFPasswordInput.tsx
@@ -1,8 +1,9 @@
-import React                                              from "react";
+import React, { ReactNode }                               from "react";
 import { useFormContext, Controller }                     from "react-hook-form";
 import { MantineSize, PasswordInput, PasswordInputProps } from "@mantine/core";
 import { useFocusWithin }                                 from "@mantine/hooks";
 import { OpenedEyeOutline, ClosedEyeOutline }             from "@inprodi/icons";
+import { InputHelpLabel, InputHelpLabelProps }            from "../InputHelpLabel";
 import styles                                             from "../styles";
 
 export interface RHFPasswordInputProps extends PasswordInputProps, React.RefAttributes<HTMLInputElement> {
@@ -34,16 +35,29 @@ export interface RHFPasswordInputProps extends PasswordInputProps, React.RefAttr
     * to change the inpt radius.To see more: [textInput](https://mantine.dev/core/text-input/)
     */
     radius?: MantineSize;
+	/**
+   * Tooltip label
+   */
+	help ?: ReactNode;
+	/**
+    * All the props to customize all the label and helper node.
+    */
+	helpLabelProps ?: Omit<InputHelpLabelProps, "label" | "helpLabel">;
 }
 export const RHFPasswordInput = ({
 	name,
+	size,
+	help,
 	label,
+	withAsterisk,
+	helpLabelProps,
 	onChange,
 	...rest
 } : RHFPasswordInputProps) => {
 	const { ref, focused } = useFocusWithin();
 	const { classes } = styles({ focused });
 	const { control } = useFormContext();
+	const showInputHelp = !!(help && label);
 
 	const getVisibilityIcon = ({ reveal, size } : { reveal : boolean; size : number }) =>{
 		return reveal ? <OpenedEyeOutline data-testid="opened-eye" fontSize={size} />
@@ -64,19 +78,32 @@ export const RHFPasswordInput = ({
 						{...field}
 						data-testid="password"
 						ref={ref}
-						error={!!error}
+						name={name}
+						size={size}
+						error={error?.message}
 						autoComplete="off"
+						withAsterisk={!showInputHelp ? withAsterisk : false}
 						onChange={handleOnChange}
-						icon={<OpenedEyeOutline fontSize={14} />}
 						visibilityToggleIcon={getVisibilityIcon}
-						label={error?.message ? error.message : label}
+						label={
+							showInputHelp ?
+								<InputHelpLabel
+									size={size}
+									label={label}
+									helpLabel={help}
+									required={withAsterisk}
+									{...helpLabelProps}
+								/>
+								:
+								label
+						}
 						classNames={{
 							icon       : classes.icon,
 							input      : classes.pswInput,
 							innerInput : classes.innerInput,
 							invalid    : classes.invalidPassword,
 							wrapper    : error && classes.invalidWrapper,
-							label      : `${classes.label} ${error && classes.labelError}`,
+							label      : classes.label,
 						}}
 						{...rest}
 					/>